Created a new app called org
- Moved everything related to organisation to the app
- If user is student display student details in clubs home page
