+++ org/admin.py
@@ -1,0 +1,7 @@
+from django.contrib import admin
+from .models import OrgProfile
+
+
+@admin.register(OrgProfile)
+class OrgProfileAdmin(admin.ModelAdmin):
+    list_display = ('name',)
